python-api/requirements.txt
Django
djangorestframework
transformers
torch
spacy
django-cors-headers